#67: 句読点をstatlc Listで管理するように変更
・ついでに句読点を使ってるメソッドで定義したstatic listを使うように変更
