+++ src/main/java/models/word/PeriodConst.java
@@ -1,0 +1,20 @@
+package models.word;
+
+import com.google.common.collect.Lists;
+
+import java.util.List;
+
+public class PeriodConst {
+    public static List<String> period = Lists.newArrayList(
+            "。"
+            , "！"
+            , "!"
+            , "？"
+            , "?"
+            , "."
+            , "．"
+    );
+}
+
+
+
